refactor: moved SequenceDataset to test
